// src/types/next-auth.d.ts
import { Role } from "@prisma/client";
import "next-auth";
import "next-auth/jwt";

declare module "next-auth" {
  interface Session {
    user: {
      id: string;
      email: string;
      name: string;
      username: string;
      avatar: string | null;
      bio: string | null;
      country: string | null;
      city: string | null;
      role: Role;
    };
  }

  interface User {
    id: string;
    username?: string;
    role?: Role;
    avatar?: string | null;
    bio?: string | null;
    country?: string | null;
    city?: string | null;
  }
}

declare module "next-auth/jwt" {
  interface JWT {
    id: string;
    username: string;
    role: Role;
    avatar: string | null;
    bio: string | null;
    country: string | null;
    city: string | null;
  }
}

export interface ApiUser {
  id: string;
  email: string;
  name: string;
  username: string;
  avatar: string | null;
  bio: string | null;
  country: string | null;
  city: string | null;
  role: Role;
}